Reorganize repository implementations
