+++ src/main/java/ua/mibal/repository/web/config/props/WebServiceProductRepositoryProps.java
@@ -1,4 +1,4 @@
-package ua.mibal.repository.impl.web.config.props;
+package ua.mibal.repository.web.config.props;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
@@ -6,8 +6,8 @@ import org.springframework.boot.context.properties.ConfigurationProperties;
  * @author Mykhailo Balakhon
  * @link <a href="mailto:mykhailo.balakhon@communify.us">mykhailo.balakhon@communify.us</a>
  */
-@ConfigurationProperties(prefix = "application.rest-client")
-public record RestClientProps(
-        int responseTimeout
+@ConfigurationProperties(prefix = "application.web-service-product-repository")
+public record WebServiceProductRepositoryProps(
+        String url
 ) {
 }
